Create separate application configuration class.
