+++ src/main/java/com/practice/mongodb/MongodbApplication.java
@@ -2,32 +2,11 @@ package com.practice.mongodb;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.web.client.RestTemplateBuilder;
-import org.springframework.context.annotation.Bean;
-import org.springframework.web.client.RestTemplate;
-
-import java.time.Duration;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
 
 @SpringBootApplication
 public class MongodbApplication {
 
 	public static void main(String[] args) {
 		SpringApplication.run(MongodbApplication.class, args);
-	}
-
-	@Bean
-	public RestTemplate restTemplate(RestTemplateBuilder builder) {
-
-		return builder
-				.setConnectTimeout(Duration.ofMillis(3000))
-				.setReadTimeout(Duration.ofMillis(3000))
-				.build();
-	}
-
-	@Bean
-	public ExecutorService executorService() {
-		return Executors.newFixedThreadPool(10);
 	}
 }
